requirements.txt
ace_tools==0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types
anyio
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.9.1
Bottleneck
cbor2==5.5.1
certifi==2023.11.17
chardet
charset-normalizer==3.3.2
colorama
contourpy
crytic-compile==0.3.5
cycler
cytoolz==0.12.2
distro
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup
fonttools
frozenlist==1.4.1
graphviz==0.20.3
h11
hexbytes==0.3.1
httpcore
httpx
idna==3.6
importlib_resources
Jinja2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver
lru-dict==1.2.0
MarkupSafe
matplotlib
matplotlib-venn==1.1.1
mkl-fft
mkl-random
mkl-service==2.4.0
multidict==6.0.4
numexpr
numpy==1.24.3
openai
packaging==23.2
pandas==2.2.2
parsimonious==0.9.0
pillow
prettytable==3.9.0
protobuf==4.25.1
py-solc-x==2.0.2
py-solidity-ast==0.1.5
pycryptodome==3.19.1
pydantic
pydantic_core
pyparsing
python-dateutil
pytz
pyunormalize==15.1.0
pywin32==306
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.16.2
scipy==1.13.1
six
slither-analyzer==0.10.0
sniffio
solc-select==1.0.4
toml==0.10.2
toolz==0.12.0
tqdm
typing_extensions
tzdata
unicodedata2
urllib3==2.1.0
wcwidth==0.2.12
web3==6.13.0
websockets==12.0
yarl==1.9.4
z3-solver==4.13.0.0
zipp
